pci: keep the upper half of a 64-bit BAR
get_bar() masked the flag bits off with ~u32(...), which zero-extends
to 64 bits and cleared the upper half of the address too. A BAR above
4 GiB came back as its low 32 bits: QEMU's arm64 virt puts an xHCI
controller at 0x80_0000_8000, and the driver went to 0x8000.

+++ kernel/pci/device.v
@@ -69,7 +69,9 @@ pub fn (dev &PCIDevice) get_bar(bar u8) PCIBar {
 	is_64_bits := is_mmio && ((bar_low >> 1) & 0b11) == 0b10
 	bar_high := if is_64_bits { dev.read[u32](reg_index + 4) } else { 0 }
 
-	base := ((u64(bar_high) << 32) | bar_low) & ~u32(if is_mmio { 0b1111 } else { 0b11 })
+	// The mask has to be 64 bits wide: ~u32 zero-extends and would drop the
+	// upper half of a BAR above 4 GiB.
+	base := ((u64(bar_high) << 32) | bar_low) & ~u64(if is_mmio { 0b1111 } else { 0b11 })
 
 	dev.write[u32](reg_index, 0xFFFFFFFF)
 	bar_size_low = dev.read[u32](reg_index)
